x/offers/types/message_accept_proposal.go
package types

import (
	errorsmod "cosmossdk.io/errors"
	sdkmath "cosmossdk.io/math"
	sdk "github.com/cosmos/cosmos-sdk/types"
	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
)

var _ sdk.Msg = &MsgAcceptProposal{}

func NewMsgAcceptProposal(creator string, id sdkmath.Uint) *MsgAcceptProposal {
	return &MsgAcceptProposal{
		Creator: creator,
		Id:      id,
	}
}

func (msg *MsgAcceptProposal) ValidateBasic() error {
	_, err := sdk.AccAddressFromBech32(msg.Creator)
	if err != nil {
		return errorsmod.Wrapf(sdkerrors.ErrInvalidAddress, "invalid creator address (%s)", err)
	}
	return nil
}
