x/offers/types/message_accept_proposal_test.go
package types

import (
	"testing"

	"bitbadgeschain/testutil/sample"

	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
)

func TestMsgAcceptProposal_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgAcceptProposal
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgAcceptProposal{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgAcceptProposal{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
